requirements.txt
azure-ai-textanalytics
nltk
matplotlib
